brain/requirements.txt
openai
asyncio
pandas
pyarrow
lightgbm
wasm-exporter
numpy
pyyaml
httpx
redis 